+++ src/api/include/vkgs/renderer.h
@@ -29,7 +29,7 @@ class VKGS_API Renderer {
    * dst: (height, width, 4) uint8 image, row-major.
    */
   RenderedImage draw(GaussianSplats splats, const float* view, const float* projection, uint32_t width, uint32_t height,
-                     uint8_t* dst);
+                     float* dst);
 
   const auto* impl() const noexcept { return impl_.get(); }
